Admin verification in generator1
Вроде как теперь бота может остановить только его создатель
