+++ TelegramBotGeneratorData/config.py
@@ -1,3 +1,4 @@
 node = {node}
 inventory_size = {inventory_size}
 BotKey = '{BotKey}'
+AdminID = {AdminId}
